extracted Countdown component
