+++ src/components/App/App.jsx
@@ -1,16 +1,12 @@
 import React, { Component } from 'react';
 
-import {
-  formatTime,
-  calculateIntervalCompletionPercentage
-} from '../../common/timeUtils';
-
 import './App.css';
+import Countdown from '../Countdown';
 
 let ipcRenderer;
 
 class App extends Component {
-  state = { timerPaused: false, timeRemaining: 0 }
+  state = { timerPaused: false, timeRemaining: 0, intervalLength: 0 }
   componentWillMount() {
     const electron = window.require('electron');
     ipcRenderer = electron.ipcRenderer;
@@ -33,23 +29,9 @@ class App extends Component {
   }
   render() {
     const { intervalLength, timerPaused, timeRemaining } = this.state;
-    const completionPercentage = calculateIntervalCompletionPercentage(
-      timeRemaining,
-      intervalLength
-    );
     return (
       <div className='App'>
-        <div className='Countdown'
-          style={{
-            background: `linear-gradient(
-              to bottom,
-              rgb(0, 51, 102),
-              rgb(0, 51, 102) ${ completionPercentage}%,
-              rgb(0, 0, 102) ${ completionPercentage}%
-            )`
-          }}>
-          <h1>{formatTime(timeRemaining)}</h1>
-        </div>
+        <Countdown timeRemaining={timeRemaining} intervalLength={intervalLength} />
         <div className='btnContainer'>
           <button className='btn'
             onClick={this.handlePlayPauseClick}>
