+++ src/components/Countdown/index.js
@@ -1,0 +1,3 @@
+import Countdown from './Countdown';
+
+export default Countdown;
